refactor(expo-polyfills): move storage into separate file, boost coverage
